Recreating the old tab stop functionality.
Version 0.3 of PHPRtfLite featured the ability to add tab stops to
ParFormat objects. This commit re-enables the feature for the latest
version of PHPRtfLite.

The code is taken directly from Denis Slaveckij's original class and
should probably be rewritten to conform to PHPRtfLite's latest
conventions. For the time being, this will re-enable old functionality
for those who require it, albeit in a hacky way.